+++ vgen/src/main/resources/io/gapi/vgen/nodejs/discovery_fragment.snip
@@ -46,7 +46,7 @@
 
     @join param : params
       @let paramField = context.getField(signatureType, param), \
-           paramValue = context.typeDefaultValue(signatureType, paramField), \
+           paramValue = context.typeDefaultValue(signatureType, paramField, method), \
            paramDescription = apiaryConfig.getDescription(signatureType.getName, param)
         {@item(paramDescription)}
         {@context.mapParamName(param)}: {@paramValue}
