+++ vgen/src/test/java/io/gapi/vgen/testdata/discoveries/nodejs/nodejs_cloudmonitoring.v2beta2.json.baseline
@@ -89,6 +89,7 @@ authFactory.getApplicationDefault(function(err, authClient) {
 
     // The project id. The value can be the numeric project ID or string-based project name.
     project: "",
+    resource: {},
     // Auth client
     auth: authClient
   };
@@ -133,6 +134,7 @@ authFactory.getApplicationDefault(function(err, authClient) {
     metric: "",
     // End of the time interval (inclusive), which is expressed as an RFC 3339 timestamp.
     youngest: "",
+    resource: {},
     // Auth client
     auth: authClient
   };
@@ -211,6 +213,7 @@ authFactory.getApplicationDefault(function(err, authClient) {
     metric: "",
     // End of the time interval (inclusive), which is expressed as an RFC 3339 timestamp.
     youngest: "",
+    resource: {},
     // Auth client
     auth: authClient
   };
