+++ vgen/src/test/java/io/gapi/vgen/testdata/discoveries/nodejs/nodejs_translate.v2.json.baseline
@@ -19,7 +19,7 @@ authFactory.getApplicationDefault(function(err, authClient) {
     // TODO: Change placeholders below to values for parameters to the 'list' method:
 
     // The text to detect
-    q: [],
+    q: "",
     // Auth client
     auth: authClient
   };
@@ -83,7 +83,7 @@ authFactory.getApplicationDefault(function(err, authClient) {
     // TODO: Change placeholders below to values for parameters to the 'list' method:
 
     // The text to translate
-    q: [],
+    q: "",
     // The target language into which the text should be translated
     target: "",
     // Auth client
